+++ src/Application/DonorTags/Commands/DeleteDonorTag/DeleteDonorTag.cs
@@ -1,6 +1,7 @@
 using VinculoBackend.Application.Common.Exceptions;
 using VinculoBackend.Application.Common.Interfaces;
 using VinculoBackend.Application.Common.Models;
+using FluentValidation.Results;
 
 namespace VinculoBackend.Application.DonorTags.Commands.DeleteDonorTag;
 
@@ -25,6 +26,18 @@ public sealed class DeleteDonorTagCommandHandler : IRequestHandler<DeleteDonorTa
         if (entity is null)
         {
             throw new Common.Exceptions.NotFoundException(nameof(DonorTags), request.Id.ToString());
+        }
+
+        var isInUse = await _context.DonorTagAssignments
+            .AsNoTracking()
+            .AnyAsync(assignment => assignment.DonorTagId == entity.Id, cancellationToken);
+
+        if (isInUse)
+        {
+            throw new Common.Exceptions.ValidationException(
+            [
+                new ValidationFailure(nameof(DeleteDonorTagCommand.Id), "Nao e possivel excluir uma tag relacionada a doadores."),
+            ]);
         }
 
         _context.DonorTags.Remove(entity);
